+++ DataBaseServer/GameFrameX.DataBase/GameDB.cs
@@ -113,7 +113,7 @@ public static class GameDb
     /// <param name="pageIndex">页码，从0开始。</param>
     /// <param name="pageSize">每页数量，默认为10。</param>
     /// <returns>符合条件的元素列表。</returns>
-    public static Task<List<TState>> FindSortDescendingAsync<TState>(Expression<Func<TState, bool>> filter, Expression<Func<TState, object>> sortExpression, long pageIndex = 0, long pageSize = 10) where TState : class, ICacheState, new()
+    public static Task<List<TState>> FindSortDescendingAsync<TState>(Expression<Func<TState, bool>> filter, Expression<Func<TState, object>> sortExpression, int pageIndex = 0, int pageSize = 10) where TState : class, ICacheState, new()
     {
         return _dbServiceImplementation.FindSortDescendingAsync<TState>(filter, sortExpression, pageIndex, pageSize);
     }
@@ -127,7 +127,7 @@ public static class GameDb
     /// <param name="pageIndex">页码，从0开始。</param>
     /// <param name="pageSize">每页数量，默认为10。</param>
     /// <returns>符合条件的元素列表。</returns>
-    public static Task<List<TState>> FindSortAscendingAsync<TState>(Expression<Func<TState, bool>> filter, Expression<Func<TState, object>> sortExpression, long pageIndex = 0, long pageSize = 10) where TState : class, ICacheState, new()
+    public static Task<List<TState>> FindSortAscendingAsync<TState>(Expression<Func<TState, bool>> filter, Expression<Func<TState, object>> sortExpression, int pageIndex = 0, int pageSize = 10) where TState : class, ICacheState, new()
     {
         return _dbServiceImplementation.FindSortAscendingAsync<TState>(filter, sortExpression, pageIndex, pageSize);
     }
